Cleaning up ScrollMagic
It's still crappy right now, but the debugging is throwing errors on the About page, and I don't want it in a broken state.

Also, as you can see by the code here, it's kind of a great big kludge right now. Not really getting some aspects of this plugin, however much I do like it.