Giving up on ScrollMagic for now
This is not working in Safari, but it works fine in Chrome. Have no idea why. This commit is a fix for an obvious issue, but it's still broken in Safari, and prevents other JS from executing on the page (like the nav), so may need to launch without the fancy effects...